requirements.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==24.3
django-json-widget==1.1.1
psycopg2-binary==2.9.10
celery==5.6.2
redis==7.1.0
drf-spectacular==0.29.0
PyYAML==6.0.3
python-dotenv==1.0.0
